pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = ""
authors = ["Jaime Junr <jaimeodairbassojuniorjaime@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "<5.0"
django-rest-framework = "^0.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
factory-boy = "^3.3.0"
pytest-django = "^4.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
